requirements.txt
arrow==0.10.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
cssselect==1.0.1
cycler==0.10.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.9.6
JPype1==0.6.2
konlpy==0.4.4
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.0.2
newspaper3k==0.2.2
nltk==3.2.4
numpy==1.13.1
oauthlib==2.0.2
olefile==0.44
pandas==0.20.3
Pillow==4.2.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-instagram==1.3.2
pytz==2017.2
PyYAML==3.12
requests==2.18.1
requests-file==1.4.2
requests-oauthlib==0.8.0
scipy==0.19.1
selenium==3.4.3
simplejson==3.11.1
six==1.10.0
tldextract==2.1.0
tweepy==3.5.0
urllib3==1.21.1
virtualenv==15.1.0
Werkzeug==0.12.2
